AspirePythonDemo/PythonApi/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
redis==5.0.1
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-exporter-otlp==1.21.0
